Task1/requirements.txt
requests==2.26.0
beautifulsoup4==4.10.0
pandas==1.3.3 
textblob==0.15.3 
wordcloud==1.8.1 
matplotlib==3.4.3 
python-pptx==0.6.19
